ftgl途中、エラー出る

+++ OpenGL_Game/src/bullet/control_bullet.cpp
@@ -94,7 +94,7 @@ void Draw() {
 	}
 //弾数を表示登録
 	char string[256];
-	sprintf(string, "bullet:%d", bullet_.size());
+	sprintf(string, "bullet:%d", (int)bullet_.size());
 	output_display::Regist(string, uColor4fv_blue, 1);
 }
 }
