+++ OpenGL_Game/src/meteo/control_meteo.cpp
@@ -89,7 +89,7 @@ void Draw() {
 	}
 	//隕石数を表示登録
 	char string[256];
-	sprintf(string, "meteo:%d", meteo_.size());
+	sprintf(string, "meteo:%d", (int)meteo_.size());
 	output_display::Regist(string, uColor4fv_blue, 1);
 }
 }
